+++ src/main/java/com/sprint/mission/discodeit/common/email/EmailSender.java
@@ -3,6 +3,7 @@ package com.sprint.mission.discodeit.common.email;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 
 // 이메일 발송 공통
@@ -12,7 +13,8 @@ public class EmailSender {
 
   private final JavaMailSender mailSender;
 
-  public void sendEmail(String to, String subject, String content) {
+  @Async
+  public void sendEmailAsync(String to, String subject, String content) {
     SimpleMailMessage message = new SimpleMailMessage();
     message.setTo(to);
     message.setSubject(subject);
