+++ src/main/java/com/sprint/mission/discodeit/config/AsyncConfig.java
@@ -1,0 +1,10 @@
+package com.sprint.mission.discodeit.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+@Configuration
+@EnableAsync
+public class AsyncConfig {
+
+}
