+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicAuthService.java
@@ -1,12 +1,10 @@
-package com.sprint.mission.discodeit.facade.auth;
+package com.sprint.mission.discodeit.service.basic;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.sprint.mission.discodeit.common.email.EmailSender;
 import com.sprint.mission.discodeit.dto.auth.response.VerifyCodeRes;
-import com.sprint.mission.discodeit.exception.CustomException;
-import com.sprint.mission.discodeit.exception.ErrorCode;
-import com.sprint.mission.discodeit.service.UserService;
+import com.sprint.mission.discodeit.service.AuthService;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
@@ -14,9 +12,8 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class EmailAuthFacade {
+public class BasicAuthService implements AuthService {
 
-  private final UserService userService;
   private final EmailSender emailSender;
   Cache<String, String> emailAuthCache = Caffeine.newBuilder()
       .expireAfterWrite(5, TimeUnit.MINUTES)  //5분 뒤 자동 삭제
@@ -31,7 +28,7 @@ public class EmailAuthFacade {
     emailAuthCache.put(email, code);
 
     // SimpleMail 보내기
-    emailSender.sendEmail(
+    emailSender.sendEmailAsync(
         email,
         "[ch-at] 이메일 인증코드",
         "이메일 인증 신청을 하지 않으신 경우 이 메일을 무시해주세요.\n" +
@@ -43,6 +40,7 @@ public class EmailAuthFacade {
   // 인증번호 검증
   public VerifyCodeRes verifyEmailCode(String email, String code) {
     String auth = emailAuthCache.getIfPresent(email);
+    System.out.println(code + " : " + auth);
     if (auth == null || !auth.equals(code)) {
       return new VerifyCodeRes(false, "인증 코드가 일치하지 않거나 만료되었습니다."); // 5분 지나면 자동 삭제
     }
